+++ src/controllers/console/Oauth2DebugController.php
@@ -7,6 +7,7 @@ use rhertogh\Yii2Oauth2Server\controllers\console\debug\Oauth2DebugConfigAction;
 
 class Oauth2DebugController extends Oauth2BaseConsoleController
 {
+    public $defaultAction = 'config';
     /**
      * @inheritDoc
      */